Adjust rules for responseHeaders matching
Apparently responseHeaders applies if *any* of the headers match, not *all* of the headers.

This more narrowly scopes matching for PDFs -- generally on URLs that return Content-Type: "application/octet-stream"

Specifically, if a URL returns a content-disposition: of attachment with *.pdf*, we'll rewrite that to content-type of PDF and to render inline.